move error class to separate file
